pyproject.toml
[project]
name = "esg-llm-platform"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "camelot-py>=1.0.9",
    "ghostscript>=0.8.1",
    "openai>=2.8.0",
    "pandas>=2.3.3",
    "pdfplumber>=0.11.8",
    "pymupdf>=1.26.6",
    "python-dotenv>=1.2.1",
    "pyyaml>=6.0.3",
]
